Comment to aid code reading; boost pylint score
When I first read the code I added some comments to help find my way around the
script and I worked on improving the Pylint score as a way to get familiar with
the script.

Hopefully these changes make it easier for other contributers to read the code.

I'd like to try integrating `tcvt` with
[`pymux`](https://github.com/jonathanslenders/pymux).